This reverts commit 7f799ea32e601af872593a3bcf85b5b9d498f5d2.

+++ objects.h
@@ -36,7 +36,7 @@ struct Zombies
     bool live; // Define quando um novo Zombie surge
     int life;
     float speed; // Velocidade do Zombie
-    int boundx; // Colisao X
+    float boundx; // Colisao X
     int boundy; // Colisao Y
 
 };
